STS-16. Added distributed keyrotation managers for mongo & jdbc
