--- extras-keyrotation/distributed-keyrotation/src/main/java/de/adorsys/keymanagement/keyrotation/api/persistence/KeyStorePersistence.java
@@ -1,9 +1,0 @@
-package de.adorsys.keymanagement.keyrotation.api.persistence;
-
-import java.security.KeyStore;
-
-public interface KeyStorePersistence {
-
-    KeyStore read();
-    void write(KeyStore keyStore);
-}
