+++ extras-keyrotation/keyrotation-core/src/main/java/de/adorsys/keymanagement/keyrotation/api/persistence/KeyStorePersistence.java
@@ -1,0 +1,7 @@
+package de.adorsys.keymanagement.keyrotation.api.persistence;
+
+public interface KeyStorePersistence {
+
+    byte[] read();
+    void write(byte[] keyStore);
+}
