+++ extras-keyrotation/keyrotation-core/src/main/java/de/adorsys/keymanagement/keyrotation/api/persistence/RotationLocker.java
