+++ extras-keyrotation/keyrotation-core/src/main/java/de/adorsys/keymanagement/keyrotation/api/services/KeyStoreManager.java
@@ -1,0 +1,6 @@
+package de.adorsys.keymanagement.keyrotation.api.services;
+
+import de.adorsys.keymanagement.api.Juggler;
+
+public interface KeyStoreManager extends Juggler {
+}
