+++ extras-keyrotation/keyrotation-core/src/main/java/de/adorsys/keymanagement/keyrotation/api/types/CommonValidities.java
